Added a full width option
Added the option to set full width. (fullwidth: true) If set to "true" then the "fullwidth" will set to the fullwidth of the ".sidr-inner" div. Also their will be created an extra div ("sidr-outer") next to the ".sidr-inner" div and the menu will get an extra class "fullwidth", this in order to style the menu in "fullwidth" mode. This will function as the closing div. Why these changes? Well sometimes you only want to focus on the menu. So that you can't scroll the content as well. Off course wen activated the CSS styles have to be different as well.